requirements.txt
torch>=2.3
torchvision>=0.18
scikit-learn>=1.3
tqdm>=4.66
pillow>=10

matplotlib>=3.8    
opencv-python>=4.8
pydicom>=2.4       
scikit-image>=0.22